fix: Used setRawText(true) and changed PromptCommand to ConfirmCommand
