+++ docs/docs/components/terminal.md
@@ -37,7 +37,7 @@ files={[
   'src/main/java/com/webforj/samples/views/terminal/commands/DateCommand.java',
   'src/main/java/com/webforj/samples/views/terminal/commands/HelpCommand.java',
   'src/main/java/com/webforj/samples/views/terminal/commands/MsgCommand.java',
-  'src/main/java/com/webforj/samples/views/terminal/commands/PromptCommand.java',
+  'src/main/java/com/webforj/samples/views/terminal/commands/ConfirmCommand.java',
   'src/main/java/com/webforj/samples/views/terminal/commands/TimeCommand.java',
   'src/main/resources/static/css/terminal/terminal-view.css',
 ]}
